refactor: controller 패키지 catalina내부로 이동
