+++ tomcat/src/test/java/org/apache/coyote/controller/LoginControllerTest.java
@@ -1,6 +1,8 @@
 package org.apache.coyote.controller;
 
 import org.apache.catalina.SessionManager;
+import org.apache.catalina.controller.AbstractController;
+import org.apache.catalina.controller.LoginController;
 import org.apache.coyote.http11.HttpHeaders;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
