+++ tomcat/src/test/java/org/apache/coyote/controller/PageControllerTest.java
@@ -1,5 +1,7 @@
 package org.apache.coyote.controller;
 
+import org.apache.catalina.controller.AbstractController;
+import org.apache.catalina.controller.PageController;
 import org.apache.coyote.http11.HttpHeaders;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
